// src/api/models/CreateVacancySchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateVacancySchema = {
    title: string;
    description: string;
    requirements: string;
    location: string;
    position_type: string;
    experience_level: string;
    expiry_date: string;
    salary_min?: (number | string | null);
    salary_max?: (number | string | null);
};